ChatWithDB/requirements.txt
groq
python-dotenv
mysql-connector-python
streamlit

